Fixing bbone_chain bug
bone chain was dependent on when they were added to the armature not on its bone hierarchy. This was fix by making sure the parent bone was the first on a list and then all the recursive children in order.
Also, there were minors tweaks on single_control.py, sorting the functions alphabetically so it easier to look up in case of changes.

+++ rig_modules/set_bone.py
@@ -3,17 +3,56 @@ from pathlib import Path
 import bpy
 import math
 
-def naming(bone: EditBone, bone_type: str) -> str:
-    '''Puts preffix to bone name. It replaces its previous one if it has.'''
-    bname = bone.name.split('-')
-    if len(bname) < 2:
-        return f'{bone_type}-{bone.name}'
-    bname[0] = bone_type
-    return '-'.join(bname)
+def assign_widget(bone: PoseBone, shape) -> any:         
 
-def parenting(bone: EditBone, parentbone: EditBone, context: Context) -> any:
-    bone.use_connect = False
-    bone.parent = parentbone
+    bone.custom_shape = shape
+    bone.custom_shape_rotation_euler[0] = math.radians(90)
+    if bone.name.startswith('strHandle') or bone.name.startswith('endHandle'):
+        bone.custom_shape_scale_xyz = 1.6, 1.6, 1.6
+        return {'FINISHED'}
+    bone.custom_shape_scale_xyz = 1.3, 1.3, 1.3
+
+
+def collection(bone: EditBone, colname: str, context: Context) -> any:
+
+    arm = context.active_object.data
+    if colname not in arm.collections: 
+        arm.collections.new(name=colname)
+    arm.collections[colname].assign(bone)
+    return {'FINISHED'}
+
+
+def bbones_prop(bone: EditBone, handle_type: str = "TANGENT") -> any:
+
+    if bone.bbone_segments == 1:
+        bone.bbone_segments = 3
+
+    bone.bbone_easein = 0.0
+    bone.bbone_easeout = 0.0
+    bone.bbone_handle_type_start = handle_type
+    bone.bbone_handle_type_end = handle_type
+
+    bone.bbone_handle_use_scale_start = True, False, True
+    bone.bbone_handle_use_scale_end = True, False, True
+
+
+def bbone_handles(bone: EditBone, bhandle: EditBone, context: Context) -> any:
+    """Adds custom bones as start/end handles for bbones."""
+
+    edit_bones = context.active_object.data.edit_bones
+    if bone.head == bhandle.head:
+        bone.bbone_custom_handle_start = edit_bones[bhandle.name]
+
+    elif bone.tail == bhandle.head:
+        bone.bbone_custom_handle_end = edit_bones[bhandle.name]
+
+
+def bone_prop(bone: EditBone, type_inscale: str = 'ALIGNED') -> any:
+    '''Set bone properties in Edit Mode.'''
+
+    bone.inherit_scale = type_inscale
+    return {"FINISHED"}
+
 
 def create(
 bone: EditBone,
@@ -23,7 +62,9 @@ bbone_size: float,
 length: float,
 context: Context,
 bone_name: str = '',
+inherir_scale: str = 'AVERAGE',
 ) -> EditBone:
+    
     if bone_name:
         bname = bone_name
     else:
@@ -37,45 +78,30 @@ bone_name: str = '',
     new_bone.bbone_x = bbone_size
     new_bone.bbone_z = new_bone.bbone_x
     new_bone.use_deform = False
-    new_bone.inherit_scale = 'AVERAGE'
+    new_bone.inherit_scale = inherir_scale
 
     return new_bone
 
 
-def bbones_prop(bone: EditBone) -> any:
-    if bone.bbone_segments == 1:
-        bone.bbone_segments = 3
+def naming(bone: EditBone, bone_type: str, sep: str = '-') -> str:
+    '''Puts preffix to bone name. It replaces its previous one if it has.'''
 
-    bone.bbone_easein = 0.0
-    bone.bbone_easeout = 0.0
-    bone.bbone_handle_type_start = "TANGENT"
-    bone.bbone_handle_type_end = "TANGENT"
-
-    bone.bbone_handle_use_scale_start[0] = True
-    bone.bbone_handle_use_scale_start[1] = False
-    bone.bbone_handle_use_scale_start[2] = True
-
-    bone.bbone_handle_use_scale_end[0] = True
-    bone.bbone_handle_use_scale_end[1] = False
-    bone.bbone_handle_use_scale_end[2] = True
+    bname = bone.name.split(sep)
+    if len(bname) < 2:
+        return f'{bone_type}-{bone.name}'
+    bname[0] = bone_type
+    return sep.join(bname)
 
 
-def bbone_handles(bone: EditBone, bhandle: EditBone, context: Context) -> any:
-    """Adds custom bones as start/end handles for bbones."""
-    edit_bones = context.active_object.data.edit_bones
-    if bone.head == bhandle.head:
-        bone.bbone_custom_handle_start = edit_bones[bhandle.name]
+def parenting(bone: EditBone, parentbone: EditBone, context: Context) -> any:
 
-    elif bone.tail == bhandle.head:
-        bone.bbone_custom_handle_end = edit_bones[bhandle.name]
+    bone.use_connect = False
+    bone.parent = parentbone
 
-def bone_prop(bone: EditBone, type_inscale: str = 'ALIGNED') -> any:
-    '''Set bone properties in Edit Mode.'''
-    bone.inherit_scale = type_inscale
-    return {"FINISHED"}
 
 def pbone_properties(bone: PoseBone) -> any:
     """Set bone properties in Pose Mode."""
+
     bone.rotation_mode = "XYZ"
 
     # if bone.bone.use_deform:
@@ -83,14 +109,20 @@ def pbone_properties(bone: PoseBone) -> any:
     #     bone.bbone_easeout = 1
     return {"FINISHED"}
 
-def collection(bone: EditBone, colname: str, context: Context) -> any:
-    arm = context.active_object.data
-    if colname not in arm.collections: 
-        arm.collections.new(name=colname)
-    arm.collections[colname].assign(bone)
-    return {'FINISHED'}
+
+def sorting(bone_chain: list) -> list:
+    '''Sorting the parent of the chain first and then all the recursive children. Not depending on order of creation.'''
+
+    bchain = [bone for bone in bone_chain if bone.parent == None]
+    if not bchain:
+        return False
+    for childbone in bchain[0].children_recursive:
+        bchain.append(childbone)
+    return bchain
+
 
 def widget(widget_name: str, context: Context) -> EditBone:
+
     current_dir = Path(__file__).parent.parent
     folder = Path('/armature_presets/widgets.blend')
     preset_folder = f'{current_dir}{folder}'
@@ -103,21 +135,14 @@ def widget(widget_name: str, context: Context) -> EditBone:
         data_to.objects = [name for name in data_from.objects if not name in bpy.data.objects.keys() and name in widget_name]
 
     widgetcoll = context.active_object.name.replace('RIG', 'WIDGET')
-    for obj in data_to.objects:
-        context.scene.collection.objects.link(obj)
-        bpy.data.collections[widgetcoll].objects.link(obj)
-        context.scene.collection.objects.unlink(obj)  
-        return obj
-    
-def assign_widget(bone: PoseBone, shape) -> any:               
-    bone.custom_shape = shape
-    bone.custom_shape_rotation_euler[0] = math.radians(90)
-    if bone.name.startswith('strHandle') or bone.name.startswith('endHandle'):
-        bone.custom_shape_scale_xyz[0] = 0.5
-        bone.custom_shape_scale_xyz[1] = bone.custom_shape_scale_xyz[0]
-        bone.custom_shape_scale_xyz[2] = bone.custom_shape_scale_xyz[0]
-        return {'FINISHED'}
-    bone.custom_shape_scale_xyz[0] = 1.3
-    bone.custom_shape_scale_xyz[1] = bone.custom_shape_scale_xyz[0]
-    bone.custom_shape_scale_xyz[2] = bone.custom_shape_scale_xyz[0]
-    
+    if widgetcoll in bpy.data.collections.keys():
+        for obj in data_to.objects:
+            context.scene.collection.objects.link(obj)
+            bpy.data.collections[widgetcoll].objects.link(obj)
+            context.scene.collection.objects.unlink(obj)  
+            return obj
+    else:
+        for obj in data_to.objects:
+            context.scene.collection.objects.link(obj)
+            return obj
+
